app.py
import streamlit as st
import cv2
import numpy as np
from keras.models import load_model
from cvzone.HandTrackingModule import HandDetector
import enchant
import pyttsx3
import os

# Initialize components
model = load_model('cnn8grps_rad1_model.h5')
hd = HandDetector(maxHands=1)
dictionary = enchant.Dict("en-US")
offset = 29

def process_frame(frame):
    # Process frame similar to original code
    hands = hd.findHands(frame, draw=False, flipType=True)
    predicted_char = None
    
    if hands[0]:
        hand = hands[0]
        x, y, w, h = hand[0]['bbox']
        
        # Image processing similar to original code
        # Return predicted character
        
    return predicted_char

def main():
    st.title("Sign Language To Text Conversion")
    
    # Initialize session state
    if 'text' not in st.session_state:
        st.session_state.text = ""
    
    # Add webcam input
    cap = cv2.VideoCapture(0)
    frame_placeholder = st.empty()
    
    # Add text display
    text_placeholder = st.empty()
    
    # Add buttons
    col1, col2 = st.columns(2)
    with col1:
        if st.button("Clear"):
            st.session_state.text = ""
    with col2:
        if st.button("Speak"):
            # Implement text-to-speech
            pass
            
    while True:
        ret, frame = cap.read()
        if not ret:
            break
            
        frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
        frame_placeholder.image(frame)
        
        predicted_char = process_frame(frame)
        if predicted_char:
            st.session_state.text += predicted_char
            
        text_placeholder.text(f"Text: {st.session_state.text}")
        
if __name__ == "__main__":
    main()