requirements.txt
absl-py==2.3.1
attrs==25.3.0
cvzone==1.6.1
mediapipe==0.10.21
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
pandas==2.2.3
pillow==11.1.0
pyttsx3==2.99
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sounddevice==0.5.2
tensorflow==2.19.0/

streamlit
opencv-python-headless
pyenchant
